Rename and move address widget folders
