--- lib/routes/receive_payment/lightning/widgets/widgets.dart
@@ -1,9 +1,0 @@
-export 'compact_qr_image.dart';
-export 'loading_or_error.dart';
-export 'particle_animations/animation_progress.dart';
-export 'particle_animations/animation_properties.dart';
-export 'particle_animations/particle_model.dart';
-export 'particle_animations/particle_painter.dart';
-export 'particle_animations/particles_animations.dart';
-export 'successful_payment.dart';
-export 'successful_payment_message.dart';
