+++ lib/routes/receive_payment/ln_address/receive_lightning_address_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:l_breez/cubit/cubit.dart';
 import 'package:l_breez/cubit/webhook/webhook_state.dart';
-import 'package:l_breez/routes/receive_payment/ln_address/widgets/destination_widget_placeholder.dart';
-import 'package:l_breez/routes/receive_payment/widgets/address_widget/destination_widget.dart';
+import 'package:l_breez/routes/receive_payment/widgets/destination_widget/destination_widget.dart';
+import 'package:l_breez/routes/receive_payment/widgets/destination_widget/destination_widget_placeholder.dart';
 import 'package:l_breez/routes/receive_payment/widgets/payment_info_message_box/payment_limits_message_box.dart';
 import 'package:l_breez/utils/exceptions.dart';
 import 'package:l_breez/widgets/single_button_bottom_bar.dart';
