+++ lib/routes/receive_payment/widgets/destination_widget/destination_widget_placeholder.dart
