+++ lib/routes/receive_payment/widgets/destination_widget/widgets/compact_qr_image.dart
