+++ lib/routes/receive_payment/widgets/destination_widget/widgets/loading_or_error.dart
