+++ lib/routes/receive_payment/widgets/destination_widget/widgets/particle_animations/animation_properties.dart
