+++ lib/routes/receive_payment/widgets/destination_widget/widgets/particle_animations/particles_animations.dart
@@ -1,9 +1,14 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:l_breez/routes/receive_payment/lightning/widgets/particle_animations/particle_model.dart';
-import 'package:l_breez/routes/receive_payment/lightning/widgets/particle_animations/particle_painter.dart';
+import 'package:l_breez/routes/receive_payment/widgets/destination_widget/widgets/particle_animations/particle_model.dart';
+import 'package:l_breez/routes/receive_payment/widgets/destination_widget/widgets/particle_animations/particle_painter.dart';
 import 'package:simple_animations/simple_animations.dart';
+
+export 'animation_progress.dart';
+export 'animation_properties.dart';
+export 'particle_model.dart';
+export 'particle_painter.dart';
 
 class Particles extends StatefulWidget {
   final int numberOfParticles;
