+++ lib/routes/receive_payment/widgets/destination_widget/widgets/widgets.dart
@@ -1,0 +1,3 @@
+export 'compact_qr_image.dart';
+export 'loading_or_error.dart';
+export 'particle_animations/particles_animations.dart';
