+++ lib/routes/receive_payment/widgets/successful_payment/successful_payment.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:l_breez/routes/receive_payment/lightning/widgets/widgets.dart';
+import 'package:l_breez/routes/receive_payment/widgets/destination_widget/widgets/widgets.dart';
+import 'package:l_breez/routes/receive_payment/widgets/successful_payment/successful_payment_message.dart';
 
 class SuccessfulPaymentRoute extends StatefulWidget {
   const SuccessfulPaymentRoute({super.key});
